upgrade libtiff to v4.0.10

+++ humble-video-captive/src/main/gnu/libtiff/csrc/port/strcasecmp.c
@@ -1,5 +1,3 @@
-/* $Id: strcasecmp.c,v 1.3 2009-01-22 20:53:07 fwarmerdam Exp $ */
-
 /*
  * Copyright (c) 1987, 1993
  *	The Regents of the University of California.  All rights reserved.
@@ -44,8 +42,8 @@ strcasecmp(const char *s1, const char *s2)
 	const unsigned char *us1 = (const unsigned char *)s1,
 			*us2 = (const unsigned char *)s2;
 
-	while (tolower(*us1) == tolower(*us2++))
+	while (tolower((int) *us1) == tolower((int) *us2++))
 		if (*us1++ == '\0')
 			return (0);
-	return (tolower(*us1) - tolower(*--us2));
+	return (tolower((int) *us1) - tolower((int) *--us2));
 }
